+++ Math/NumberTheory/ArithmeticFunctions/Standard.hs
@@ -10,12 +10,13 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Math.NumberTheory.ArithmeticFunctions.Standard
-  ( -- * Multiplicative functions
-    multiplicative
-  , divisors, divisorsA
+  ( -- * List divisors
+    divisors, divisorsA
   , divisorsList, divisorsListA
   , divisorsSmall, divisorsSmallA
   , divisorsTo, divisorsToA
+    -- * Multiplicative functions
+  , multiplicative
   , divisorCount, tau, tauA
   , sigma, sigmaA
   , totient, totientA
@@ -40,7 +41,6 @@ import qualified Data.IntSet as IS
 import Data.Maybe
 import Data.Set (Set)
 import qualified Data.Set as S
-import qualified Data.List as L
 import Data.Semigroup
 
 import Math.NumberTheory.ArithmeticFunctions.Class
@@ -99,17 +99,17 @@ divisorsHelperSmall p 1 = IS.singleton p
 divisorsHelperSmall p a = IS.fromDistinctAscList $ p : p * p : map (p ^) [3 .. wordToInt a]
 {-# INLINE divisorsHelperSmall #-}
 
--- | See `divisorsToA`
-divisorsTo :: (UniqueFactorisation n, Ord n, Integral n) => n -> n -> Set n
+-- | See 'divisorsToA'.
+divisorsTo :: (UniqueFactorisation n, Integral n) => n -> n -> Set n
 divisorsTo to = runFunction (divisorsToA to)
 
--- | The set of all (positive) divisors below `to` (inclusive)
-divisorsToA :: (UniqueFactorisation n, Ord n, Integral n) => n -> ArithmeticFunction n (Set n)
+-- | The set of all (positive) divisors up to an inclusive bound.
+divisorsToA :: (UniqueFactorisation n, Integral n) => n -> ArithmeticFunction n (Set n)
 divisorsToA to = ArithmeticFunction f unwrap
   where f p k = BoundedSetProduct (\bound -> divisorsToHelper bound (unPrime p) k)
         unwrap (BoundedSetProduct res) = if 1 <= to then S.insert 1 (res to) else res to
 
--- | Generate at most `a` powers of `p` up to bound `b` (inclusive)
+-- | Generate at most @a@ powers of @p@ up to an inclusive bound @b@.
 divisorsToHelper :: (Ord n, Num n) => n -> n -> Word -> Set n
 divisorsToHelper _ _ 0 = S.empty
 divisorsToHelper b p 1 = if p <= b then S.singleton p else S.empty
@@ -350,11 +350,11 @@ instance Num a => Monoid (ListProduct a) where
   mappend = (<>)
 
 -- Represent as a Reader monad
-newtype BoundedSetProduct a = BoundedSetProduct { getBoundedSetProduct :: a -> Set a }
+newtype BoundedSetProduct a = BoundedSetProduct { _getBoundedSetProduct :: a -> Set a }
 
 takeWhileLE :: Ord a => a -> Set a -> Set a
 takeWhileLE b xs = if m then S.insert b ls else ls
-  where (ls, m, _) = S.splitMember b xs 
+  where (ls, m, _) = S.splitMember b xs
 
 instance (Ord a, Num a) => Semigroup (BoundedSetProduct a) where
   BoundedSetProduct f1 <> BoundedSetProduct f2 = BoundedSetProduct f